move OnionClientAuth extension functions to separate file
