+++ library/kmp-tor-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/common/clientauth/OnionClientAuthPrivateKey_B64_X25519.kt
@@ -22,6 +22,7 @@ import io.matthewnelson.component.encoding.base32.encodeBase32
 import io.matthewnelson.kmp.tor.common.address.OnionAddressV3
 import io.matthewnelson.kmp.tor.common.annotation.InternalTorApi
 import io.matthewnelson.kmp.tor.common.util.TorStrings.REDACTED
+import io.matthewnelson.kmp.tor.common.util.descriptorString
 import kotlin.jvm.JvmInline
 import kotlin.jvm.JvmStatic
