+++ library/kmp-tor-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/common/clientauth/OnionClientAuthPublicKey_B64_X25519.kt
@@ -18,6 +18,7 @@ package io.matthewnelson.kmp.tor.common.clientauth
 import io.matthewnelson.component.base64.decodeBase64ToArray
 import io.matthewnelson.component.encoding.base32.Base32
 import io.matthewnelson.component.encoding.base32.encodeBase32
+import io.matthewnelson.kmp.tor.common.util.descriptorString
 import kotlin.jvm.JvmInline
 
 /**
